requirements.txt
# AgenticRag Dependencies
# Run: pip install -r requirements.txt

# ----------------------------
# UI
# ----------------------------
streamlit

# ----------------------------
# LLM & Groq
# ----------------------------
groq
python-dotenv

# ----------------------------
# Vector Store & Embeddings
# ----------------------------
faiss-cpu
sentence-transformers

# ----------------------------
# PDF Processing
# ----------------------------
pymupdf
pdfplumber
pillow
pytesseract

# ----------------------------
# Data
# ----------------------------
numpy
pandas
